+++ diagrams/src/diagrams_builder.js
@@ -1,17 +1,19 @@
 (function(root, factory) {
     if (typeof define === 'function' && define.amd) {
         // AMD. Register as an anonymous module.
-        define(['../vendor/railroad-diagrams', 'chevrotain'], factory)
+        // TODO: remove dependency to Chevrotain
+        define(['../vendor/railroad-diagrams'], factory)
     } else if (typeof module === 'object' && module.exports) {
         // Node. Does not work with strict CommonJS, but
         // only CommonJS-like environments that support module.exports,
         // like Node.
-        module.exports = factory(require('../vendor/railroad-diagrams'), require('chevrotain'))
+        // TODO: remove dependency to Chevrotain
+        module.exports = factory(require('../vendor/railroad-diagrams'))
     } else {
         // Browser globals (root is window)
-        root.diagrams_builder = factory(root.railroad, root.chevrotain)
+        root.diagrams_builder = factory(root.railroad)
     }
-}(this, function(railroad, chevrotain) {
+}(this, function(railroad) {
 
     var Diagram = railroad.Diagram
     var Sequence = railroad.Sequence
@@ -23,7 +25,7 @@
     var NonTerminal = railroad.NonTerminal
 
     /**
-     * @param {chevrotain.gast.Rule[]} topRules
+     * @param {chevrotain.gast.ISerializedGast} topRules
      *
      * @returns {string} - The htmlText that will render the diagrams
      */
@@ -46,45 +48,23 @@
     }
 
     /**
-     * @param {chevrotain.gast.Terminal} prod
+     * @param {chevrotain.gast.ISerializedTerminal} prod
      * @param {string} topRuleName
      * @param {string} dslRuleName
      *
      * @return {RailRoadDiagram.Terminal}
      */
-    function createTerminalFromGastProd(prod, topRuleName, dslRuleName) {
-
-        var pattern = prod.terminalType.PATTERN
+    function createTerminalFromSerializedGast(prod, topRuleName, dslRuleName) {
         // PATTERN static property will not exist when using custom lexers (hand built or other lexer generators)
-        var toolTipTitle = pattern ? pattern.source : undefined
-        return railroad.Terminal(chevrotain.tokenLabel(prod.terminalType),
+        var toolTipTitle = prod.pattern
+        return railroad.Terminal(prod.label,
             undefined,
             toolTipTitle,
             prod.occurrenceInParent,
             topRuleName,
             dslRuleName,
-            chevrotain.tokenName(prod.terminalType)
+            prod.name
         )
-    }
-
-    /**
-     * @param {function} tokenConstructor
-     * @param {number} occurrenceInParent
-     * @param {string} topRuleName
-     * @param {string} dslRuleName
-     *
-     * @return {RailRoadDiagram.Terminal}
-     */
-    function createTerminalFromToken(tokenConstructor, occurrenceInParent, topRuleName, dslRuleName) {
-        var result = Terminal(chevrotain.tokenLabel(tokenConstructor),
-            undefined,
-            // PATTERN static property will not exist when using custom lexers (hand built or other lexer generators)
-            tokenConstructor.PATTERN ? tokenConstructor.PATTERN.source : undefined,
-            occurrenceInParent,
-            topRuleName,
-            dslRuleName,
-            chevrotain.tokenName(tokenConstructor))
-        return result
     }
 
     /**
@@ -97,20 +77,20 @@
      */
     function convertProductionToDiagram(prod, topRuleName) {
 
-        if (prod instanceof chevrotain.gast.NonTerminal) {
+        if (prod.type === "NonTerminal") {
             // must handle NonTerminal separately from the other AbstractProductions as we do not want to expand the subDefinition
             // of a reference and cause infinite loops
-            return NonTerminal(prod.nonTerminalName, undefined, prod.occurrenceInParent, topRuleName)
+            return NonTerminal(getNonTerminalName(prod), undefined, prod.occurrenceInParent, topRuleName)
         }
-        else if (!(prod instanceof chevrotain.gast.Terminal)) {
+        else if (prod.type !== "Terminal") {
             var subDiagrams = definitionsToSubDiagrams(prod.definition, topRuleName)
-            if (prod instanceof chevrotain.gast.Rule) {
+            if (prod.type === "Rule") {
                 return Diagram.apply(this, subDiagrams)
             }
-            else if (prod instanceof chevrotain.gast.Flat) {
+            else if (prod.type === "Flat") {
                 return Sequence.apply(this, subDiagrams)
             }
-            else if (prod instanceof chevrotain.gast.Option) {
+            else if (prod.type === "Option") {
                 if (subDiagrams.length > 1) {
                     return Optional(Sequence.apply(this, subDiagrams))
                 }
@@ -121,7 +101,7 @@
                     throw Error("Empty Optional production, OOPS!")
                 }
             }
-            else if (prod instanceof chevrotain.gast.Repetition) {
+            else if (prod.type === "Repetition") {
                 if (subDiagrams.length > 1) {
                     return ZeroOrMore(Sequence.apply(this, subDiagrams))
                 }
@@ -132,11 +112,11 @@
                     throw Error("Empty Optional production, OOPS!")
                 }
             }
-            else if (prod instanceof chevrotain.gast.Alternation) {
+            else if (prod.type === "Alternation") {
                 // todo: what does the first argument of choice (the index 0 means?)
                 return Choice.apply(this, [0].concat(subDiagrams))
             }
-            else if (prod instanceof chevrotain.gast.RepetitionMandatory) {
+            else if (prod.type === "RepetitionMandatory") {
                 if (subDiagrams.length > 1) {
                     return OneOrMore(Sequence.apply(this, subDiagrams))
                 }
@@ -147,34 +127,49 @@
                     throw Error("Empty Optional production, OOPS!")
                 }
             }
-            else if (prod instanceof chevrotain.gast.RepetitionWithSeparator) {
+            else if (prod.type === "RepetitionWithSeparator") {
                 if (subDiagrams.length > 0) {
                     // MANY_SEP(separator, definition) === (definition (separator definition)*)?
                     return Optional(Sequence.apply(this, subDiagrams.concat(
                         [ZeroOrMore(Sequence.apply(this,
-                            [createTerminalFromToken(prod.separator, prod.occurrenceInParent, topRuleName, "many_sep")].concat(subDiagrams)))])))
+                            [createTerminalFromSerializedGast(
+                                prod.separator,
+                                topRuleName,
+                                "many_sep")].concat(subDiagrams)))])))
                 }
                 else {
                     throw Error("Empty Optional production, OOPS!")
                 }
             }
-            else if (prod instanceof chevrotain.gast.RepetitionMandatoryWithSeparator) {
+            else if (prod.type === "RepetitionMandatoryWithSeparator") {
                 if (subDiagrams.length > 0) {
                     // AT_LEAST_ONE_SEP(separator, definition) === definition (separator definition)*
                     return Sequence.apply(this, subDiagrams.concat(
                         [ZeroOrMore(Sequence.apply(this,
-                            [createTerminalFromToken(prod.separator, prod.occurrenceInParent, topRuleName, "at_least_one_sep")].concat(subDiagrams)))]))
+                            [createTerminalFromSerializedGast(
+                                prod.separator,
+                                topRuleName,
+                                "at_least_one_sep")].concat(
+                                subDiagrams)))]))
                 }
                 else {
                     throw Error("Empty Optional production, OOPS!")
                 }
             }
         }
-        else if (prod instanceof chevrotain.gast.Terminal) {
-            return createTerminalFromGastProd(prod, topRuleName, "consume")
+        else if (prod.type === "Terminal") {
+            return createTerminalFromSerializedGast(prod, topRuleName, "consume")
         }
         else {
             throw Error("non exhaustive match")
+        }
+    }
+
+    function getNonTerminalName(prod) {
+        if (prod.nonTerminalName !== undefined) {
+            return prod.nonTerminalName
+        } else {
+            return prod.name
         }
     }