+++ diagrams/src/main.js
@@ -14,8 +14,13 @@
 }(this, function(builder, behavior) {
     return {
         drawDiagramsFromParserInstance: function(parserInstanceToDraw, targetDiv) {
-            var topRules = parserInstanceToDraw.getGAstProductions().values()
+            var topRules = parserInstanceToDraw.getSerializedGastProductions()
             targetDiv.innerHTML = builder.buildSyntaxDiagramsText(topRules);
+            behavior.initDiagramsBehavior()
+        },
+
+        drawDiagramsFromSerializedGrammar: function(serializedGrammar, targetDiv) {
+            targetDiv.innerHTML = builder.buildSyntaxDiagramsText(serializedGrammar);
             behavior.initDiagramsBehavior()
         }
     };